+++ Examples/Map+Positioning/Sources/ViewControllers/CameraViewController.swift
@@ -24,11 +24,12 @@ class CameraViewController: UIViewController {
     
     private let stateListener = SerialDisposable()
     private let disposeBag = DisposeBag()
+    private var arView: ARView!
     
     override func viewDidLoad() {
         super.viewDidLoad()
         
-        let arView = ARView(frame: view.frame)
+        arView = ARView(frame: view.frame)
         arView.autoresizingMask = [.flexibleWidth, .flexibleHeight]
         view.insertSubview(arView, at: 0)
         arView.session = session
@@ -68,6 +69,20 @@ class CameraViewController: UIViewController {
                 dismiss(animated: true)
             })
     }
+    
+    // FIXME: workaround to avoid automatic stopping of the session due to iOS 18 changes.
+    // before if you set your custom session, you have to start and stop it manually.
+    // Now even if it's a custom session - it will be automatically stopped when view is hidden.
+    // And it's not even possible to set nil to reset the session. So here we set new session just to remove control from our own one.
+    // ARSession does not report any changes in state or errors when it happens, that's why we consider this behaviour as a bug.
+    // so you can't differentiate this behaviour is tracking lost or a bug. because last reported state is normal, but position is not updated
+    override func viewWillDisappear(_ animated: Bool) {
+        if #available(iOS 18.0, *) {
+            arView.session = ARSession()
+        }
+        super.viewWillDisappear(animated)
+    }
+    // end
     
     @IBAction func stopScan() {
         vpsLocationSource.stopScan()